extract common parts from cppcheck_createrules.py
* move common parts to utils_createrules.py
* fix pylint warnings and apply autopep8

preparation for #1536
